Implement chord_stack (#2241)
This commit implements a chord stack to keep track of the chord modes
that were activated. When exiting a mode, a higher-level chord with
the mode attribute is searched for. If found, it is activated.
Otherwise the root keybindings are restored.
A new command is introduced: cmd_ungrab_all_chords.
Unlike before, cmd_ungrab_chord does not directly go to the root mode.
This can now be explicitly done with cmd_ungrab_all_chords.