feature parity for `CheckUpdates` widget.
The `CheckUpdates` widget argument `custom_command` now is allowed to be
a tuple or list `(str, int)` indicating the command and the number of
lines to subtract. This guarantees feature parity with the built-in
configurations for various package managers.
Passing a string is still supported, the default value of 0 lines to
subtract will be used.

This commit moves all of the decision logic about `custom_command` from
`_check_updates` to `__init__`.