Add a *using `git`* how-to to the docs.
This commit adds a practical guide to some tasks that are frequently
needed when contributing. This includes creating and checkout out
branches, committing, pushing, pulling, amending, force-pushing,
rebasing, rewording and squashing. It is not meant to be a beginner
guide to `git`, but rather a to show how to use the knowledge acquired
when contributing to qtile.